+++ internal/impl/confluent/processor_schema_registry_encode.go
@@ -29,7 +29,22 @@ Encodes messages automatically from schemas obtains from a [Confluent Schema Reg
 
 If a message fails to encode under the schema then it will remain unchanged and the error can be caught using error handling methods outlined [here](/docs/configuration/error_handling).
 
-Currently only Avro schemas are supported.`).
+Currently only Avro schemas are supported.
+
+### Avro JSON Format
+
+By default this processor expects documents formatted as [Avro JSON](https://avro.apache.org/docs/current/spec.html#json_encoding) when encoding Avro schemas. In this format the value of a union is encoded in JSON as follows:
+
+- if its type is ` + "`null`, then it is encoded as a JSON `null`" + `;
+- otherwise it is encoded as a JSON object with one name/value pair whose name is the type's name and whose value is the recursively encoded value. For Avro's named types (record, fixed or enum) the user-specified name is used, for other types the type name is used.
+
+For example, the union schema ` + "`[\"null\",\"string\",\"Foo\"]`, where `Foo`" + ` is a record name, would encode:
+
+- ` + "`null` as `null`" + `;
+- the string ` + "`\"a\"` as `{\"string\": \"a\"}`" + `; and
+- a ` + "`Foo` instance as `{\"Foo\": {...}}`, where `{...}` indicates the JSON encoding of a `Foo`" + ` instance.
+
+However, it is possible to instead consume documents in raw JSON format (that match the schema) by setting the field ` + "[`avro_raw_json`](#avro_raw_json) to `true`" + `.`).
 		Field(service.NewStringField("url").Description("The base URL of the schema registry service.")).
 		Field(service.NewInterpolatedStringField("subject").Description("The schema subject to derive schemas from.").
 			Example("foo").
@@ -39,6 +54,9 @@ Currently only Avro schemas are supported.`).
 			Default("10m").
 			Example("60s").
 			Example("1h")).
+		Field(service.NewBoolField("avro_raw_json").
+			Description("Whether messages encoded in Avro format should be parsed as raw JSON documents rather than [Avro JSON](https://avro.apache.org/docs/current/spec.html#json_encoding).").
+			Advanced().Default(false)).
 		Field(service.NewTLSField("tls")).
 		Version("3.58.0")
 }
@@ -60,6 +78,7 @@ func init() {
 type schemaRegistryEncoder struct {
 	client             *http.Client
 	subject            *service.InterpolatedString
+	avroRawJSON        bool
 	schemaRefreshAfter time.Duration
 
 	schemaServerURL *url.URL
@@ -82,6 +101,10 @@ func newSchemaRegistryEncoderFromConfig(conf *service.ParsedConfig, logger *serv
 	if err != nil {
 		return nil, err
 	}
+	avroRawJSON, err := conf.FieldBool("avro_raw_json")
+	if err != nil {
+		return nil, err
+	}
 	refreshPeriodStr, err := conf.FieldString("refresh_period")
 	if err != nil {
 		return nil, err
@@ -98,13 +121,14 @@ func newSchemaRegistryEncoderFromConfig(conf *service.ParsedConfig, logger *serv
 	if err != nil {
 		return nil, err
 	}
-	return newSchemaRegistryEncoder(urlStr, tlsConf, subject, refreshPeriod, refreshTicker, logger)
+	return newSchemaRegistryEncoder(urlStr, tlsConf, subject, avroRawJSON, refreshPeriod, refreshTicker, logger)
 }
 
 func newSchemaRegistryEncoder(
 	urlStr string,
 	tlsConf *tls.Config,
 	subject *service.InterpolatedString,
+	avroRawJSON bool,
 	schemaRefreshAfter, schemaRefreshTicker time.Duration,
 	logger *service.Logger,
 ) (*schemaRegistryEncoder, error) {
@@ -116,6 +140,7 @@ func newSchemaRegistryEncoder(
 	s := &schemaRegistryEncoder{
 		schemaServerURL:    u,
 		subject:            subject,
+		avroRawJSON:        avroRawJSON,
 		schemaRefreshAfter: schemaRefreshAfter,
 		schemas:            map[string]*cachedSchemaEncoder{},
 		shutSig:            shutdown.NewSignaller(),
@@ -321,20 +346,31 @@ func (s *schemaRegistryEncoder) getLatestEncoder(subject string) (schemaEncoder,
 	}
 
 	var codec *goavro.Codec
-	if codec, err = goavro.NewCodec(resPayload.Schema); err != nil {
+	if codec, err = goavro.NewCodecForStandardJSON(resPayload.Schema); err != nil {
 		s.logger.Errorf("failed to parse response for schema subject '%v': %v", subject, err)
 		return nil, 0, err
 	}
 
 	return func(m *service.Message) error {
-		datum, err := m.AsStructured()
-		if err != nil {
+		var datum interface{}
+		if s.avroRawJSON {
+			b, err := m.AsBytes()
+			if err != nil {
+				return err
+			}
+
+			if datum, _, err = codec.NativeFromTextual(b); err != nil {
+				return err
+			}
+		} else if datum, err = m.AsStructured(); err != nil {
 			return err
 		}
+
 		binary, err := codec.BinaryFromNative(nil, datum)
 		if err != nil {
 			return err
 		}
+
 		m.SetBytes(binary)
 		return nil
 	}, resPayload.ID, nil